do some package cleanup before we start deriving our cloed source cli from here
