+++ cmd/keystore_commands.go
@@ -3,12 +3,9 @@ package cmd
 import (
 	"errors"
 	"fmt"
-	"strings"
 
 	pb "github.com/RTradeLtd/TxPB/v3/go"
 	"github.com/RTradeLtd/go-temporalx-sdk/client"
-	crypto "github.com/libp2p/go-libp2p-core/crypto"
-	"github.com/libp2p/go-libp2p-core/peer"
 	au "github.com/logrusorgru/aurora"
 	"github.com/urfave/cli/v2"
 )
@@ -37,30 +34,7 @@ func keystoreCreate() *cli.Command {
 			if err != nil {
 				return err
 			}
-			var pk crypto.PrivKey
-			switch strings.ToLower(c.String("key.type")) {
-			case "rsa":
-				pk, _, err = crypto.GenerateKeyPair(
-					crypto.RSA,
-					4096,
-				)
-			case "ed25519":
-				pk, _, err = crypto.GenerateKeyPair(
-					crypto.Ed25519,
-					256,
-				)
-			case "ecdsa":
-				pk, _, err = crypto.GenerateKeyPair(
-					crypto.ECDSA,
-					256,
-				)
-			default:
-				return errors.New("key.type flag is empty or contains incorrect value")
-			}
-			if err != nil {
-				return err
-			}
-			pid, err := peer.IDFromPrivateKey(pk)
+			pk, pid, err := createIPFSKey(c.String("key.type"), c.Int("key.size"))
 			if err != nil {
 				return err
 			}
@@ -80,6 +54,6 @@ func keystoreCreate() *cli.Command {
 			})
 			return err
 		},
-		Flags: []cli.Flag{keyName(), keyType()},
+		Flags: []cli.Flag{keyName(), keyType(), keySize(), mnemonicFlag()},
 	}
 }
