+++ go.mod
@@ -4,6 +4,7 @@ go 1.14
 
 require (
 	github.com/RTradeLtd/TxPB/v3 v3.2.0
+	github.com/bonedaddy/nemonify v0.0.0-20200401092711-1e348e4c9242
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
 	github.com/google/go-cmp v0.3.1 // indirect
 	github.com/libp2p/go-libp2p-core v0.5.0
